+++ NXDNGateway/KenwoodNetwork.cpp
@@ -33,7 +33,7 @@ const unsigned char BIT_MASK_TABLE[] = { 0x80U, 0x40U, 0x20U, 0x10U, 0x08U, 0x04
 
 const unsigned int BUFFER_LENGTH = 200U;
 
-CKenwoodNetwork::CKenwoodNetwork(unsigned int localPort, const std::string& rptAddress, unsigned int rptPort, bool debug) :
+CKenwoodNetwork::CKenwoodNetwork(unsigned short localPort, const std::string& rptAddress, unsigned short rptPort, bool debug) :
 m_rtpSocket(localPort + 0U),
 m_rtcpSocket(localPort + 1U),
 m_rtcpAddr(),
